+++ languages/zh_cn/operations.json
@@ -51,11 +51,11 @@
   "subscribeForum": "关注专业",
   "visitThread": "阅读文章",
   "postToThread": "回复文章",
-  "moveThread": "移动文章",
+  "moveThread": "将文章移动至回收站",
   "collectThread": "收藏文章",
   "delCollectThread": "取消文章收藏",
   "digestThread": "将文章设为精选文章",
-  "toppedThread": "置顶文章",
+  "toppedThread": "专业置顶",
   "switchInPersonalForum": "在专栏加精、置顶、隐藏",
   "visitUserCard": "访问用户名片",
   "bannedUser": "封禁用户",
@@ -234,13 +234,13 @@
   "modifyXsfSettings": "修改<后台管理-设置-学术分>",
   "modifyKcbSettings": "修改科创币设置",
   "modifyRoleUsers": "将用户从某个角色中移除",
-  "modifyOtherPosts": "修改别人的post",
+  "modifyOtherPosts": "修改别人的文章回复",
   "moveThreadToRecycle": "移动文章到回收站",
   "moveDraft": "退休文章",
   "displayRecycleMarkThreads": "显示被退回的文章",
   "displayDisabledPosts": "显示被屏蔽的post",
   "unDigestThread": "取消将文章设为文章优秀",
-  "unToppedThread": "取消将文章置顶",
+  "unToppedThread": "取消专业置顶",
   "recommendPost": "推荐回复",
   "unRecommendPost": "取消推荐回复",
   "displayPostHideHistories": "显示post被屏蔽的历史",
@@ -285,8 +285,8 @@
   "modifyPasswordByEmail": "通过邮箱修改密码",
   "findPasswordVerifyEmail": "找回密码中的验证邮件",
   "visitUserTransaction": "查看用户交易信息",
-  "digestPost": "回复加精",
-  "unDigestPost": "取消回复加精",
+  "digestPost": "文章回复加精",
+  "unDigestPost": "取消文章回复加精",
   "closeThread": "关闭主题",
   "openThread": "开放主题",
   "rollbackPost": "post回滚",
