+++ nkcModules/logger.js
@@ -16,7 +16,7 @@ module.exports = async (ctx) => {
     processTime,
     uid: ctx.data.user? ctx.data.user.uid : 'visitor'
   };
-  const operation = await OperationModel.findOne({_id: ctx.data.operationId});
+  // const operation = await OperationModel.findOne({_id: ctx.data.operationId});
   const {operationId} = ctx.data;
   const d = {
     url: log.path,
@@ -32,9 +32,13 @@ module.exports = async (ctx) => {
   if(ctx.logIt) {
     d.from = ctx.req.headers.referer;
     d.address = ctx.address;
+    let operationName = operationId || '';
+    if(ctx.state.lang) {
+      operationName = ctx.state.lang('operations', operationId) || '未知操作';
+    }
     if (ctx.error) {
       console.error(
-        `${moment().format('YYYY/MM/DD HH:mm:ss').grey} ${(' ' + global.NKC.processId + ' ').grey} ${' Error '.bgRed} ${log.uid.bgCyan} ${log.method.black.bgYellow} ${log.path.bgBlue} <${processTime.green}ms> ${String(log.status).red} ${(ctx.state.lang('operations', operationId) || '未知操作').grey}`
+        `${moment().format('YYYY/MM/DD HH:mm:ss').grey} ${(' ' + global.NKC.processId + ' ').grey} ${' Error '.bgRed} ${log.uid.bgCyan} ${log.method.black.bgYellow} ${log.path.bgBlue} <${processTime.green}ms> ${String(log.status).red} ${operationName.grey}`
       );
       d.error = ctx.error;
       global.NKC.io.of('/console').NKC.webMessage(d);
@@ -42,7 +46,7 @@ module.exports = async (ctx) => {
         console.error(log.error);
     } else {
       console.log(
-        `${moment().format('YYYY/MM/DD HH:mm:ss').grey} ${(' ' + global.NKC.processId + ' ').grey} ${' Info '.bgGreen} ${log.uid.bgCyan} ${log.method.black.bgYellow} ${log.path.bgBlue} <${processTime.green}ms> ${String(log.status).green} ${(ctx.state.lang('operations', operationId)||'未知操作').grey}`
+        `${moment().format('YYYY/MM/DD HH:mm:ss').grey} ${(' ' + global.NKC.processId + ' ').grey} ${' Info '.bgGreen} ${log.uid.bgCyan} ${log.method.black.bgYellow} ${log.path.bgBlue} <${processTime.green}ms> ${String(log.status).green} ${operationName.grey}`
       );
       global.NKC.io.of('/console').NKC.webMessage(d);
     }
