+++ pages/experimental/nav.pug
@@ -112,8 +112,6 @@
           li
             a(href='/e/settings/operation') 操作类型
           li
-            a(href='/e/settings/role') 用户角色
-          li
             a(href='/e/settings/user') 用户信息
           li
             a(href='/e/settings/grade') 用户等级
