+++ pages/interface_profile_info_section2.pug
@@ -16,7 +16,7 @@
     span 总主题 #{user.threadCount||0} 帖
     span 总回复 #{user.postCount||0} 楼
 
-    -var listCertText = user.navbarDesc.cs
+    -var listCertText = user.info.certsName
     span 拥有证书：#{listCertText}
     br
     span 注册于 #{dateTimeString(user.toc)}
