+++ pages/module_forum_settings.pug
@@ -18,13 +18,16 @@ if data.type !== 'category' && data.type !== 'merge'
     a.list-group-item(href=`/f/${data.forum.fid}/settings/category`)
       span.fa.fa-user-circle &nbsp;&nbsp;
       span 分类设置
-    a.list-group-item(href=`/f/${data.forum.fid}/settings/permission`)
-      span.fa.fa-user-circle &nbsp;&nbsp;
-      span 权限设置
-    a.list-group-item(href=`/f/${data.forum.fid}/settings/merge`)
-      span.fa.fa-user-circle &nbsp;&nbsp;
-      span 合并设置
-button.btn.btn-danger.btn-block(onclick=`deleteForum("${data.forum.fid}")`) 删除当前专业
+    if permission('visitForumPermissionSettings')  
+      a.list-group-item(href=`/f/${data.forum.fid}/settings/permission`)
+        span.fa.fa-user-circle &nbsp;&nbsp;
+        span 权限设置
+    if permission('visitForumMergeSettings')    
+      a.list-group-item(href=`/f/${data.forum.fid}/settings/merge`)
+        span.fa.fa-user-circle &nbsp;&nbsp;
+        span 合并设置
+if permission('deleteForum')        
+  button.btn.btn-danger.btn-block(onclick=`deleteForum("${data.forum.fid}")`) 删除当前专业
 h4 当前专业
 .text-center.h2
   a(href=`/f/${data.forum.fid}`)= data.forum.displayName
