+++ routes/post/credit.js
@@ -14,9 +14,15 @@ router
 		if(targetUser.uid === user.uid) ctx.throw(403, '不允许给自己加减学术分');
 		const thread = await post.extendThread();
     const forums = await thread.extendForums(['mainForums', 'minorForums']);
-    for(const f of forums) {
-      await f.ensureModeratorsPermission(data);
+    let isModerator = ctx.permission('superModerator');
+    if(!isModerator) {
+      for(const f of forums) {
+        isModerator = await f.isModerator(user);
+        if(isModerator) break;
+      }
     }
+    if(!isModerator) ctx.throw(400, '权限不足');
+    data.isModerator = isModerator;
 		if(thread.disabled || thread.disabled) {
 			ctx.throw(403,'无法给禁用的帖子或回复评学术分');
 		}
@@ -74,9 +80,15 @@ router
     const thread = await post.extendThread();
     const targetUser = await db.UserModel.findOnly({uid: post.uid});
     const forums = await thread.extendForums(['mainForums', 'minorForums']);
-    for(const f of forums) {
-      await f.ensureModeratorsPermission(data);
+    let isModerator = ctx.permission('superModerator');
+    if(!isModerator) {
+      for(const f of forums) {
+        isModerator = await f.isModerator(user);
+        if(isModerator) break;
+      }
     }
+    if(!isModerator) ctx.throw(400, '权限不足');
+    data.isModerator = isModerator;
     if(reason.length < 2) ctx.throw(400, '撤销原因写的太少啦~');
     const oldXsf = targetUser.xsf;
     targetUser.xsf -= record.num;
