+++ routes/thread/topped.js
@@ -7,15 +7,15 @@ router
 		const {user} = data;
     const thread = await db.ThreadModel.findOnly({tid});
     const forums = await thread.extendForums(['mainForums', 'minorForums']);
-    let isModerator;
-    for(const f of forums) {
-      isModerator = await f.isModerator(user?user.uid: '');
-      if(isModerator) break;
+    let isModerator = ctx.permission('superModerator');
+    if(!isModerator) {
+      for(const f of forums) {
+        isModerator = await f.isModerator(user?user.uid: '');
+        if(isModerator) break;
+      }
     }
 		if(!isModerator) {
-			if(!data.userOperationsId.includes('toppedThread')) {
-				ctx.throw(403, '您没有权限给该文章设置置顶');
-			}
+			ctx.throw(403, '您没有权限给该文章设置置顶');
 		}
 		if(thread.topped) ctx.throw(400, '该文章已被设置置顶');
 		await thread.update({topped: true});
@@ -27,14 +27,15 @@ router
 		const {user} = data;
 		const thread = await db.ThreadModel.findOnly({tid});
     const forums = await thread.extendForums(['mainForums', 'minorForums']);
-    for(const f of forums) {
-      isModerator = await f.isModerator(user?user.uid: '');
-      if(isModerator) break;
+    let isModerator = ctx.permission('superModerator');
+    if(!isModerator) {
+      for(const f of forums) {
+        isModerator = await f.isModerator(user?user.uid: '');
+        if(isModerator) break;
+      }
     }
 		if(!isModerator) {
-			if(!data.userOperationsId.includes('toppedThread')) {
-				ctx.throw(403, '您没有权限取消给该文章设置置顶');
-			}
+			ctx.throw(403, '您没有权限取消给该文章设置置顶');
 		}
 		if(!thread.topped) ctx.throw(400, '该文章未被设置置顶');
 		await thread.update({topped: false});
