+++ settings/operations/forum.js
@@ -1,49 +1,49 @@
 module.exports = {
-	GET: 'visitForumsCategory',
-	POST: 'addForum',
+	GET: 'visitForumsCategory', // 查看全部专业
+	POST: 'addForum', // 添加专业
 	PARAMETER: {
-		GET: 'visitForumHome',
-		POST: 'postToForum',
-		DELETE: 'deleteForum',
+		GET: 'visitForumHome', // 查看专业主页
+		POST: 'postToForum', // 在专业中发表文章
+		DELETE: 'deleteForum', // 删除专业
 		latest: {
-			GET: 'visitForumLatest'
+			GET: 'visitForumLatest' // 查看专业最新文章列表
 		},
 		home: {
-			GET: 'visitForumHome'
+			GET: 'visitForumHome' // 查看专业主页
 		},
 		visitors: {
-			GET: 'viewForumVisitors'
+			GET: 'viewForumVisitors' // 查看来访者
 		},
 		followers: {
-			GET: 'viewForumFollowers'
+			GET: 'viewForumFollowers' // 查看关注者
 		},
 		settings: {
-			GET: 'visitForumInfoSettings',
+			GET: 'visitForumInfoSettings', // 查看专业基本信息设置
 			info: {
-				GET: 'visitForumInfoSettings',
-				PATCH: 'modifyForumInfo'
+				GET: 'visitForumInfoSettings', // 查看专业基本信息设置
+				PATCH: 'modifyForumInfo' // 修改专业基本信息设置
 			},
 			merge: {
-				GET: 'visitForumMergeSettings',
-				PATCH: 'modifyMergeSettings',
+				GET: 'visitForumMergeSettings', // 查看专业合并
+				PATCH: 'modifyMergeSettings', // 修改专业合并设置
 			},
 			image: {
-				GET: 'visitForumImageSettings',
+				GET: 'visitForumImageSettings', // 查看专业头像设置
 			},
 			category: {
-				GET: 'visitForumCategorySettings',
-				PATCH: 'modifyForumCategory',
-				POST: 'addForumCategory',
-				DELETE: 'removeForumCategory'
+				GET: 'visitForumCategorySettings', // 查看专业分类设置
+				PATCH: 'modifyForumCategory', // 修改专业分类设置
+				POST: 'addForumCategory', // 添加专业下的文章分类
+				DELETE: 'removeForumCategory' // 删除专业下的文章分类
 			},
 			permission: {
-				GET: 'visitForumPermissionSettings',
-				PATCH: 'modifyForumPermission'
+				GET: 'visitForumPermissionSettings', // 查看专业权限设置
+				PATCH: 'modifyForumPermission' // 修改专业权限设置
 			}
 		},
 		subscribe: {
-			DELETE: 'unSubscribeForum',
-			POST: 'subscribeForum'
+			DELETE: 'unSubscribeForum', // 用户取消关注专业
+			POST: 'subscribeForum' // 用户关注专业
 		}
 	}
 };
